Refactor AdvancedOptions to separate class
Moved the AdvancedOptions class from ServerCfg.cs to its own file AdvancedOptions.cs for better separation of concerns and maintainability. Updated ServerCfg to use the new AdvancedOptions property and adjusted related logic to reference the new class.
