+++ FASTER/Models/ServerCfg.cs
@@ -773,77 +773,17 @@ namespace FASTER.Models
             { ServerCfgContent = ProcessFile(); }
         }
 
-    [Serializable]
-    public class AdvancedOptions : INotifyPropertyChanged
-    {
-        private bool   logObjectNotFound       = true;       // logging enabled
-        private bool   skipDescriptionParsing  = false;      // Parse description.ext
-        private bool   ignoreMissionLoadErrors = false;      // Do not ignore errors
-        private int    queueSizeLogG           = 0;          // If a specific players message queue is larger than <Value> and '#monitor' is running, dump his messages to a logfile for analysis
-
-            public bool LogObjectNotFound
-        {
-            get => logObjectNotFound;
-            set
-            {
-                logObjectNotFound = value;
-                RaisePropertyChanged("LogObjectNotFound");
-            }
-        }
-
-            public bool SkipDescriptionParsing
-        {
-            get => skipDescriptionParsing;
-            set
-            {
-                skipDescriptionParsing = value;
-                RaisePropertyChanged("SkipDescriptionParsing");
-            }
-        }
-
-            public bool IgnoreMissionLoadErrors
-        {
-            get => ignoreMissionLoadErrors;
-            set
-            {
-                ignoreMissionLoadErrors = value;
-                RaisePropertyChanged("IgnoreMissionLoadErrors");
-            }
-        }
-
-            public int QueueSizeLogG
-        {
-            get => queueSizeLogG;
-            set
-            {
-                queueSizeLogG = value;
-                RaisePropertyChanged("QueueSizeLogG");
-            }
-        }
-		
-		private string advancedOptionsContent;
-
-        public string AdvancedOptionsContent
-        {
-            get => advancedOptionsContent;
-            set
-            {
-                advancedOptionsContent = value;
-                RaisePropertyChanged("AdvancedOptionsContent");
-            }
-        }
-
-        public AdvancedOptions()
-        {
-            if(string.IsNullOrWhiteSpace(AdvancedOptionsContent))
-            { AdvancedOptionsContent = ProcessFile(); }
-        }
-
         private void Item_PropertyChanged(object sender, PropertyChangedEventArgs e)
         {
             RaisePropertyChanged("MissionChecked");
             RaisePropertyChanged("MissionContentOverride");
         }
+
+        // Grab Advanced output
+        public AdvancedOptions AdvancedOptions { get; set; } = new();
+
+        // Optional convenience property to expose the raw content
+        public string AdvancedOptionsContent => AdvancedOptions.AdvancedOptionsContent;
 
         public string ProcessFile()
         {
@@ -867,6 +807,8 @@ namespace FASTER.Models
                 { MissionContentOverride = compiledMission; }
             }
 
+            public string GenerateCfgFile()
+            {
             string output = "//\r\n"
                           + "// server.cfg\r\n"
                           + "//\r\n"
@@ -913,10 +855,6 @@ namespace FASTER.Models
                           + $"persistent = {persistent};\t\t\t\t// If 1, missions still run on even after the last player disconnected.\r\n"
                           + $"timeStampFormat = \"{timeStampFormat}\";\t\t// Set the timestamp format used on each report line in server-side RPT file. Possible values are \"none\" (default),\"short\",\"full\".\r\n"
                           + $"BattlEye = {battlEye};\t\t\t\t// Server to use BattlEye system\r\n"
-                          + $"queueSizeLogG = {QueueSizeLogG};\t\t\t// If a specific players message queue is larger than Value number and #monitor is running, dump his messages to a logfile for analysis \r\n"
-                          + $"LogObjectNotFound = {LogObjectNotFound};\t\t// When false to skip logging 'Server: Object not found messages'.\r\n"
-                          + $"SkipDescriptionParsing = {SkipDescriptionParsing};\t\t// When true to skip parsing of description.ext/mission.sqm. Will show pbo filename instead of configured missionName. OverviewText and such won't work, but loading the mission list is a lot faster when there are many missions \r\n"
-                          + $"ignoreMissionLoadErrors = {IgnoreMissionLoadErrors};\t\t// When set to true, the mission will load no matter the amount of loading errors. If set to false, the server will abort mission's loading and return to mission selection.\r\n"
                           + $"forcedDifficulty = {forcedDifficulty};\t\t\t// Forced difficulty (Recruit, Regular, Veteran, Custom)\r\n"
                           + "\r\n"
                           + "// TIMEOUTS\r\n"
@@ -955,9 +893,10 @@ namespace FASTER.Models
                           + "// HEADLESS CLIENT\r\n"
                           + $"{(headlessClientEnabled && !headlessClients.Exists(string.IsNullOrWhiteSpace) ? $"headlessClients[] =  { "{\n\t\"" + string.Join("\",\n\t \"", headlessClients) + "\"\n}" };\r\n" : "")}"
                           + $"{(headlessClientEnabled && !localClient.Exists(string.IsNullOrWhiteSpace)? $"localClient[] =  { "{\n\t\"" + string.Join("\",\n\t \"", localClient) + "\"\n}" };" : "")}";
+                   output + = AdvancedOptionsContent;
             return output;
+            }
         }
-
         public event PropertyChangedEventHandler PropertyChanged;
 
         private void RaisePropertyChanged(string property)
@@ -1012,5 +951,4 @@ namespace FASTER.Models
             PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(property));
         }
     }
-  }
 }
